refactor: Add data context to clean App.tsx
